Removed SDPeakTSF
TSpectrumFit changes interface with ROOT-6 (doubles instead of floats), which
makes compatibility with both ROOT-5 and ROOT-6 complicated. As the peak model
of TSpectrumFit is different anyhow and the fit limited to a fixed peak width
over the whole range, it doesn't seem worthwile to keep it (for now). Maybe
reintroduce later.